move types to the new lib folder
