+++ src/components/modals/Save/Save.tsx
@@ -2,7 +2,7 @@ import { Story } from '@data-story-org/core';
 import { observer } from 'mobx-react-lite';
 import React, { FC } from 'react';
 import { Store } from '../../..//store';
-import { SerializedReactDiagram } from '../../../types';
+import { SerializedReactDiagram } from '../../../lib/types';
 import { BaseModalHeader } from '../BaseModalHeader';
 import { BaseStoryWidgetModal } from '../BaseStoryModal';
 import { SaveStoryI } from '../BaseStoryModal/SaveStoryI';
