+++ src/components/pages/Inspector/InspectorTable/InspectorTable.tsx
@@ -2,7 +2,7 @@ import React, { FC, useState } from 'react';
 import { Feature } from '@data-story-org/core';
 import { InspectorTableBody } from './InspectorTableBody';
 import { InspectorTableHeading } from './InspectorTableHeading';
-import { InspectorMode } from '../../../../types';
+import { InspectorMode } from '../../../../lib/types';
 
 interface Props {
   features: Feature[];
