+++ src/components/widgets/DataStory/Story.tsx
@@ -2,7 +2,7 @@ import { observer } from 'mobx-react-lite';
 import React, { FC, useState } from 'react';
 import Modal from 'react-modal';
 import { Store } from '../../../store';
-import { GenericStory } from '../../../types';
+import { GenericStory } from '../../../lib/types';
 import { modalStyle } from '../../../utils/modalStyle';
 import { StoryWidgetModal } from '../../modals/StoryWidget';
 import { DataStoryWidgetActions } from './StoryActions';
