+++ src/diagram/models/DiagramModel.ts
@@ -3,7 +3,7 @@ import {
   DiagramModel as DefaultDiagramModel,
 } from '@projectstorm/react-diagrams';
 import { NodeModel } from './NodeModel';
-import { SerializedReactDiagram } from '../../types';
+import { SerializedReactDiagram } from '../../lib/types';
 import { version as VERSION } from '../../utils';
 import { stringify } from 'flatted';
