+++ src/lib/types/BootPayload.ts
