+++ src/lib/types/Inspector.ts
