+++ src/lib/types/SerializedLinkModel.ts
