+++ src/lib/types/SerializedNodeModel.ts
