+++ src/lib/types/SerializedPortModel.ts
