+++ src/lib/types/SerializedReactDiagram.ts
