+++ src/lib/types/globalWindow.ts
