+++ src/sections/Header/Header.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react';
 import { ActiveStory } from './ActiveStory';
-import * as globalWindow from '../../types/globalWindow';
+import * as globalWindow from '../../lib/types/globalWindow';
 import { Store } from '../../store';
 
 interface Props {
